Breaking out lexer and updating Makefile
